refactor: ✨ move netlify functions folder
